OME-Zarr Multichannel
