+++ modules/HortaTracer/src/main/java/org/janelia/horta/NeuronTracerTopComponent.java
@@ -335,12 +335,9 @@ public final class NeuronTracerTopComponent extends TopComponent
         });
 
         OmeZarrVolumeActor.getInstance().setVolumeState(volumeState);
-        OmeZarrVolumeActor.getInstance().getDynamicTileUpdateObservable().addObserver(new Observer() {
-            @Override
-            public void update(Observable o, Object arg) {
-                getNeuronMPRenderer().setIntensityBufferDirty();
-                redrawNow();
-            }
+        OmeZarrVolumeActor.getInstance().addDynamicTileUpdateObservable((o, arg) -> {
+            getNeuronMPRenderer().setIntensityBufferDirty();
+            redrawNow();
         });
 
         neuronMPRenderer = setUpActors();
